ADD: Adding new code from Ken K to calculate sea surface temperatures from sky and surface irt measurements.  Can utilize existing code from aeri2irt, so reorganized codes a bit to make it easier.  Also saw that different constants were being used so reconciled that aspect.
